+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.4.1 (TBD, 2020)
 * Bug Fixes
     * Fixed bug where setting `always_show_hint=True` did not show a hint when completing `Settables`
+    * Fixed bug in editor detection logic on Linux systems that do not have `which`
 
 ## 1.4.0 (November 11, 2020)
 * Bug Fixes