+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 * Bug Fixes
     * Fixed bug where setting `always_show_hint=True` did not show a hint when completing `Settables`
     * Fixed bug in editor detection logic on Linux systems that do not have `which`
+* Enhancements
+    * Added `silent_startup_script` option to `cmd2.Cmd.__init__()`. If `True`, then the startup script's
+      output will be suppressed. Anything written to stderr will still display.
 
 ## 1.4.0 (November 11, 2020)
 * Bug Fixes