+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.4.1 (TBD, 2020)
+## 1.5.0 (January 31, 2021)
 * Bug Fixes
     * Fixed bug where setting `always_show_hint=True` did not show a hint when completing `Settables`
     * Fixed bug in editor detection logic on Linux systems that do not have `which`
@@ -7,7 +7,9 @@
 * Enhancements
     * Added `silent_startup_script` option to `cmd2.Cmd.__init__()`. If `True`, then the startup script's
       output will be suppressed. Anything written to stderr will still display.
-    * cmd2 now uses pyreadline3 when running Python 3.8 or greater on Windows  
+    * cmd2 now uses pyreadline3 when running Python 3.8 or greater on Windows
+* Notes
+    * This is the last release planned to support Python 3.5
 
 ## 1.4.0 (November 11, 2020)
 * Bug Fixes